Fix unnecessary u64 cast flagged by Clippy (#5)
`get_result.meta.size` is already a `u64`, so the `as u64` cast is
redundant and triggers a Clippy warning that fails the CI Check workflow.

+++ server/src/store.rs
@@ -197,7 +197,7 @@ impl Store {
 
         let get_result = result?;
         let version = Self::extract_version(&get_result);
-        let object_size = get_result.meta.size as u64;
+        let object_size = get_result.meta.size;
 
         let data = get_result
             .bytes()
